Decouple wrapV1 logic from main.ts#wrap
This cl decouples the implementation logic from wrapping V1 functions
and places it into a separate directory.

I went back and re-exported the same exports from v1 back out to main.ts
to reduce disruption from existing usage.
